refactor(examples): 提取示例文件公共逻辑到共享辅助模块
创建 utils/connection-helpers.ts 共享辅助模块，提取以下通用逻辑：
- createDefaultCallbacks(): 创建默认连接回调函数
- printTools(): 打印工具列表
- printToolsWithSchema(): 打印工具详细信息（含参数结构）
- printConnectionStatus(): 打印连接状态
- printToolResult(): 打印工具调用结果
- runExample(): MCPConnection 示例运行框架
- runManagerExample(): MCPManager 示例运行框架
- setupManagerEventListeners(): 设置管理器事件监听器
- printToolsByServer(): 按服务分组打印工具列表
- printAllTools(): 打印所有可用工具

重构以下示例文件使用共享辅助模块：
- stdio.ts: 从 149 行减少到 89 行
- http.ts: 从 121 行减少到 75 行
- sse.ts: 从 121 行减少到 75 行
- streamable-http.ts: 从 154 行减少到 111 行
- call-tool.ts: 从 231 行减少到 159 行
- multi-mcp-manage.ts: 从 224 行减少到 139 行

效果：
- 减少 ~200+ 行重复代码，降至仅 8 行（0.8%）
- 符合 DRY 原则
- 降低维护成本
- 提升示例代码一致性

Fixes #3075
