+++ packages/mcp-core/examples/multi-mcp-manage.ts
@@ -31,6 +31,12 @@
  */
 
 import { MCPManager } from "@xiaozhi-client/mcp-core";
+import {
+  setupManagerEventListeners,
+  runManagerExample,
+  printToolsByServer,
+  printAllTools,
+} from "./utils/connection-helpers";
 
 /**
  * 主函数
@@ -38,34 +44,13 @@ import { MCPManager } from "@xiaozhi-client/mcp-core";
 async function main(): Promise<void> {
   console.log("=== MCPManager 多服务管理示例 ===\n");
 
-  // 1. 创建管理器
+  // 创建管理器
   const manager = new MCPManager();
 
-  // 2. 配置事件监听
-  manager.on("connect", () => {
-    console.log("🔄 开始连接所有服务...");
-  });
+  // 配置事件监听（使用通用辅助函数）
+  setupManagerEventListeners(manager);
 
-  manager.on("connected", ({ serverName, tools }) => {
-    console.log(`✅ 服务 ${serverName} 已连接`);
-    console.log(`   发现 ${tools.length} 个工具`);
-    console.log();
-  });
-
-  manager.on("error", ({ serverName, error }) => {
-    console.error(`❌ 服务 ${serverName} 出错: ${error.message}`);
-  });
-
-  manager.on("disconnected", ({ serverName, reason }) => {
-    console.log(`👋 服务 ${serverName} 已断开`);
-    console.log(`   原因: ${reason || "正常关闭"}`);
-  });
-
-  manager.on("disconnect", () => {
-    console.log("🔄 所有服务已断开");
-  });
-
-  // 3. 添加服务配置
+  // 添加服务配置
   console.log("配置服务:");
   console.log("  1. calculator - 计算器服务");
   console.log("     提供: 数学表达式计算功能");
@@ -94,54 +79,13 @@ async function main(): Promise<void> {
     args: ["-y", "@xiaozhi-client/datetime-mcp"],
   });
 
-  try {
-    // 4. 连接所有服务
-    console.log("正在连接到服务...");
-    console.log("(首次运行可能需要下载 MCP 服务包，请耐心等待...)");
-    console.log();
-
-    await manager.connect();
-
-    // 5. 获取所有已连接的服务
-    const connectedServers = manager.getConnectedServerNames();
-    console.log("已连接的服务:");
-    for (const serverName of connectedServers) {
-      console.log(`  - ${serverName}`);
-    }
-    console.log();
-
-    // 6. 分别列出每个服务的工具
-    console.log("各服务的工具列表:");
-    console.log();
-
+  // 使用通用框架运行示例
+  await runManagerExample(manager, async () => {
+    // 分别列出每个服务的工具
     const allTools = manager.listTools();
+    printToolsByServer(allTools);
 
-    // 按服务分组工具
-    const toolsByServer: Record<string, typeof allTools> = {};
-    for (const tool of allTools) {
-      if (!toolsByServer[tool.serverName]) {
-        toolsByServer[tool.serverName] = [];
-      }
-      toolsByServer[tool.serverName].push(tool);
-    }
-
-    // 打印每个服务的工具
-    for (const [serverName, tools] of Object.entries(toolsByServer)) {
-      console.log(`【${serverName}】`);
-      console.log(`  工具数量: ${tools.length}`);
-      console.log("  工具列表:");
-      for (const tool of tools) {
-        console.log(`    - ${tool.name}`);
-        if (tool.description) {
-          console.log(`      描述: ${tool.description}`);
-        }
-      }
-      console.log();
-    }
-
-    // 7. 调用示例工具
-
-    // 调用 calculator 服务的工具
+    // 调用示例工具
     console.log("调用 calculator 服务:");
     console.log("  工具: calculator");
     console.log("  参数: { expression: '12 * 3 + 4' }");
@@ -186,34 +130,9 @@ async function main(): Promise<void> {
     }
     console.log();
 
-    // 8. 查询服务状态
-    console.log("服务状态:");
-    const allStatus = manager.getAllServerStatus();
-    for (const [serverName, status] of Object.entries(allStatus)) {
-      console.log(`  【${serverName}】`);
-      console.log(`    已连接: ${status.connected ? "是" : "否"}`);
-      console.log(`    工具数: ${status.toolCount}`);
-    }
-    console.log();
-
-    // 9. 列出所有可用工具（跨服务）
-    console.log("所有可用工具（跨服务）:");
-    for (const tool of allTools) {
-      console.log(`  ${tool.serverName}/${tool.name}`);
-    }
-    console.log();
-  } catch (error) {
-    console.error("执行过程中出错:");
-    if (error instanceof Error) {
-      console.error(`  ${error.message}`);
-    }
-  } finally {
-    // 10. 断开所有连接
-    console.log("正在断开所有连接...");
-    await manager.disconnect();
-    console.log();
-    console.log("=== 示例结束 ===");
-  }
+    // 列出所有可用工具（跨服务）
+    printAllTools(allTools);
+  });
 }
 
 // 运行主函数
