+++ packages/mcp-core/examples/stdio.ts
@@ -33,6 +33,10 @@
  */
 
 import { MCPConnection } from "@xiaozhi-client/mcp-core";
+import {
+  createDefaultCallbacks,
+  runExample,
+} from "./utils/connection-helpers";
 
 /**
  * 主函数
@@ -40,7 +44,7 @@ import { MCPConnection } from "@xiaozhi-client/mcp-core";
 async function main(): Promise<void> {
   console.log("=== stdio MCP 连接示例 ===\n");
 
-  // 2. 创建连接实例
+  // 创建连接实例
   const connection = new MCPConnection(
     "calculator",
     {
@@ -48,48 +52,12 @@ async function main(): Promise<void> {
       command: "npx",
       args: ["-y", "@xiaozhi-client/calculator-mcp"],
     },
-    {
-      // 连接成功回调
-      onConnected: (data) => {
-        console.log(`✅ 服务 ${data.serviceName} 已连接`);
-        console.log(`   发现 ${data.tools.length} 个工具`);
-        console.log();
-      },
-
-      // 连接失败回调
-      onConnectionFailed: (data) => {
-        console.error(`❌ 服务 ${data.serviceName} 连接失败`);
-        console.error(`   错误: ${data.error.message}`);
-      },
-
-      // 断开连接回调
-      onDisconnected: (data) => {
-        console.log(`👋 服务 ${data.serviceName} 已断开`);
-        console.log(`   原因: ${data.reason || "正常关闭"}`);
-      },
-    }
+    createDefaultCallbacks(),
   );
 
-  try {
-    // 3. 建立连接
-    console.log("正在连接到服务...");
-    console.log("(首次运行可能需要下载 MCP 服务包，请耐心等待...)");
-    console.log();
-
-    await connection.connect();
-
-    // 4. 获取工具列表
-    const tools = connection.getTools();
-    console.log("可用工具:");
-    for (const tool of tools) {
-      console.log(`  - ${tool.name}`);
-      if (tool.description) {
-        console.log(`    描述: ${tool.description}`);
-      }
-    }
-    console.log();
-
-    // 5. 调用工具
+  // 使用通用框架运行示例
+  await runExample(connection, async () => {
+    // 调用工具
     console.log("调用工具: calculator");
     console.log("参数: { expression: '1 + 1' }");
 
@@ -106,7 +74,7 @@ async function main(): Promise<void> {
     }
     console.log();
 
-    // 6. 再调用一次，展示更多计算
+    // 再调用一次，展示更多计算
     console.log("再调用一次: calculator");
     console.log("参数: { expression: '2 * 3 + 4' }");
 
@@ -120,25 +88,7 @@ async function main(): Promise<void> {
       console.log(`  ${result2.content[0].text}`);
     }
     console.log();
-
-    // 7. 检查连接状态
-    console.log("连接状态:");
-    console.log(`  是否已连接: ${connection.isConnected()}`);
-    const status = connection.getStatus();
-    console.log(`  状态: ${status.connectionState}`);
-  } catch (error) {
-    console.error("执行过程中出错:");
-    if (error instanceof Error) {
-      console.error(`  ${error.message}`);
-    }
-  } finally {
-    // 8. 断开连接
-    console.log();
-    console.log("正在断开连接...");
-    await connection.disconnect();
-    console.log();
-    console.log("=== 示例结束 ===");
-  }
+  });
 }
 
 // 运行主函数
